@@ -6,35 +6,33 @@
   "main": "gulpfile.js",
   "scripts": {
     "dev": "gulp dev",
-    "build": "gulp packProduction && electron-builder -mw"
+    "build": "gulp packProduction && electron-builder -mw",
+    "build-w": "gulp packProduction && electron-builder -w"
   },
   "author": "yoshiyuki mizogami",
   "license": "MIT",
   "devDependencies": {
-    "@types/node": "^14.17.19",
-    "css-loader": "^6.3.0",
+    "@types/node": "^17.0.8",
+    "css-loader": "^6.5.1",
     "del": "^6.0.0",
-    "electron": "^15.0.0-beta.7",
-    "electron-builder": "^22.11.7",
-    "eslint-plugin-vue": "^7.18.0",
+    "electron": "^17.0.0-beta.1",
+    "electron-builder": "^22.14.5",
+    "esbuild": "^0.14.11",
+    "esbuild-loader": "^2.18.0",
+    "eslint": "^8.6.0",
+    "eslint-plugin-vue": "^8.2.0",
     "gulp": "^4.0.2",
-    "mini-css-extract-plugin": "^2.3.0",
-    "stylus": "^0.55.0",
-    "stylus-loader": "^6.1.0",
-    "ts-loader": "^9.2.6",
+    "mini-css-extract-plugin": "^2.4.6",
+    "stylus": "^0.56.0",
+    "stylus-loader": "^6.2.0",
     "typescript": "^4.4.3",
-    "vue": "^2.6.12",
+    "vue": "^3.2.26",
     "vue-devtools": "^5.1.4",
-    "vue-loader": "^15.9.8",
+    "vue-loader": "^17.0.0",
     "vue-style-loader": "^4.1.3",
     "vue-template-compiler": "^2.6.14",
-    "vuex": "^3.6.2",
-    "webpack": "^5.54.0"
-  },
-  "dependencies": {
-    "@electron/remote": "^2.0.1",
-    "dexie": "^3.0.3",
-    "vuedraggable": "^2.24.3"
+    "webpack": "^5.65.0",
+    "webpack-cli": "^4.9.1"
   },
   "build": {
     "appId": "yoshiyuki.mizogami.mylnext",
@@ -68,5 +66,10 @@
       "installerIcon": "app/imgs/icon.ico",
       "uninstallerIcon": "app/imgs/icon.ico"
     }
+  },
+  "dependencies": {
+    "dexie": "^3.2.0",
+    "mitt": "^3.0.0",
+    "vuedraggable": "^4.1.0"
   }
 }
