@@ -1,24 +1,20 @@
-import Vue from 'vue'
-import remote from '@electron/remote'
-import {clipboard, shell} from 'electron'
+import {clipboard, shell, ipcRenderer} from 'electron'
 import {join} from 'path'
 import Category from '../ts/models/category'
 import Item from '../ts/models/item'
 import getFileInfo from './utils/get-file-info'
 import MylDB from './utils/scheme'
-import hub from './event-hub'
-import Vuex, {Store} from 'vuex'
-import langSwitch from './lib/lang-swicher'
+import langSwitchFn, {LangUI} from './lib/lang-swicher'
 import switchTheme from './lib/switch-theme'
 import {Themes} from './consts'
 import Config from './models/config';
 import globals from './globals'
 import {writeFile, existsSync, readFile} from 'fs';
 import eventHub from './event-hub';
+import { reactive, nextTick, watch } from 'vue'
 
 const {DEFAULT_JSON_NAME} = globals
-Vue.use(Vuex)
-const thisWindow = remote.getCurrentWindow()
+
 
 export enum Langs{
   EN = 'en',
@@ -26,218 +22,221 @@ export enum Langs{
 }
 const config = new Config()
 const db = new MylDB()
-const isUrl = /^https?:\/\//
-const storeData = {
-  state:{
+export const isUrl = /^https?:\/\//
+export const state = reactive({
     version:globals.VERSION,
     themes:Themes,
-    categories:[],
-    items:[],
+    categories:[] as Category[],
+    items:[] as any[],
     showNewCategoryDialog:false,
-    selectedCategory:null,
+    selectedCategory:null as null|Category,
     config,
     configRaw:null as any,
     loading:false,
-    ui:{},
+    ui:{} as LangUI,
     sortMode:false,
     dragItem:null as object | unknown
-  },
-  mutations:{
-    switchSortMode(state:any){
-      state.sortMode = !state.sortMode
-    },
-    setNewCategoryDialog(state:any, tf:boolean){
-      state.showNewCategoryDialog = tf
-    },
-    setLoading(state:any, tf:boolean){
-      state.loading = tf
-    },
-    setSelectedCategory(state:any, c:Category){
-      state.selectedCategory = c
-    },
-    toggleAOT(state:any){
-      state.config.aot = !state.config.aot
-    },
-    setDragItem(state:any, item:object){
-      state.dragItem = item
-    }
-  },
-  actions:{
-    async moveItem({state}, destCategory:any){
-      if(destCategory === state.selectedCategory){
-        return
+  }
+)
+export function switchSortMode(){
+  state.sortMode = !state.sortMode
+}
+export function setNewCategoryDialog(tf:boolean){
+  state.showNewCategoryDialog = tf
+}
+export function setLoading(tf:boolean){
+  state.loading = tf
+}
+export function setSelectedCategory(c:Category){
+  state.selectedCategory = c
+}
+export function toggleAOT(){
+  state.config.aot = !state.config.aot
+}
+export function setDragItem(item:object){
+  state.dragItem = item
+}
+export async function moveItem(destCategory:any){
+  if(destCategory === state.selectedCategory){
+    return
+  }
+  const {dragItem} = state
+  state.dragItem = null
+  const ind = state.items.indexOf(dragItem)
+  state.items.splice(ind, 1)
+  await db.moveItem(dragItem, destCategory.id)
+}
+export async function init(){
+  loadConfig()
+  state.categories = await db.getCategories()
+  state.selectedCategory = state.categories[0]
+  state.ui = await langSwitchFn(state.config.lang)
+  ipcCommunicate('showWindow')
+}
+function ipcCommunicate(channel:string, data:any = undefined){
+  ipcRenderer.send(channel, data)
+}
+async function ipcHandleCommunicate(channel:string, data:any = undefined){
+  return await ipcRenderer.invoke(channel, data)
+}
+
+export async function loadConfig(){
+  state.configRaw = await db.loadConfig()
+  state.config = state.configRaw.config
+}
+export async function saveConfig(){
+  db.saveConfig(state.configRaw)
+}
+export async function langSwitch(lang:string){
+  state.config.lang = lang as any
+  state.ui = langSwitchFn(lang as any)
+  saveConfig()
+}
+export async function selectTheme(theme:string){
+  state.config.theme = theme
+  saveConfig()
+  switchTheme(theme)
+}
+export function updateItemsOrder(newOrders:Array<Item>):void{
+  state.items = newOrders
+  newOrders.forEach((o, i)=>{
+    o.sort = i
+    db.items.update(o.id, {'sort':o.sort})
+  })
+}
+export function updateCategoriesOrder(newOrders:Array<Category>):void{
+  state.categories = newOrders
+  newOrders.forEach((o, i)=>{
+    o.sort = i
+    db.categories.update(o.id as any, {'sort':o.sort})
+  })
+}
+export function updateCategoryName(cate:Category){
+  db.categories.update(cate.id as any, {name:cate.name})
+}
+export async function addNewCategory( categoryName:string){
+  const cate:Category = await db.addCategory(categoryName)
+  state.categories.push(cate)
+  state.selectedCategory = cate
+  nextTick(()=>eventHub.emit('adjust'))
+}
+export async function removeCategory(cate:Category){
+  await db.removeCategory(cate)
+  const removeIndex = state.categories.findIndex(c=> c === cate)
+  state.categories.splice(removeIndex, 1)
+  if(state.selectedCategory === cate){
+    state.items = []
+  }
+  nextTick(()=>eventHub.emit('adjust'))
+}
+export async function getItems(cateId:number){
+  const items = await db.getItems(cateId)
+  state.items = items
+  nextTick(()=>eventHub.emit('adjust'))
+}
+export function openHP(){
+  shell.openExternal(globals.HP_URL)
+}
+export async function removeItem(item:Item){
+  await db.removeItem(item)
+  const removeIndex = state.items.findIndex((i:any)=> i === item)
+  state.items.splice(removeIndex, 1)
+  nextTick(()=>eventHub.emit('adjust'))
+}
+export async function addFile({filepath} :{filepath:string}){
+  const cateId = state.selectedCategory?.id
+  const items = state.items
+  const fileInfo = await getFileInfo(filepath)
+  fileInfo.cateId = cateId
+  const fileItem = await db.addItem(fileInfo)
+  items.push(fileItem)
+  nextTick(()=>eventHub.emit('adjust'))
+}
+export async function addUrl({url, name}:{url:string, name:string}){
+  if(!isUrl.test(url)){
+    eventHub.emit('notify', state.ui.INVALID_URL)
+    return
+  }
+  const urlItem = await db.addUrlItem(name, url, state.selectedCategory?.id as any)
+  state.items.push(urlItem)
+  nextTick(()=>eventHub.emit('adjust'))
+}
+export async function copyItemPath(item:Item){
+  clipboard.writeText(item.path)
+}
+export async function showItemDetail(item:Item){
+  eventHub.emit('show-item-detail', item)
+}
+export async function updateItem(item:Item){
+  db.items.update(item.id, item)
+}
+//TODO
+function getDesktopPath(){
+  return ''
+}
+export async function importJson(){
+  const desktopPath = await ipcHandleCommunicate('getDesktopPath')
+  const targetJsonFiles = await ipcHandleCommunicate('showOpenDialog', {
+    title:'Select Myl save data json.',
+    defaultPath:join(getDesktopPath(), DEFAULT_JSON_NAME),
+    filters:[
+      {
+        name:'Json',
+        extensions:['json']
       }
-      const {dragItem} = state
-      state.dragItem = null
-      const ind = state.items.indexOf(dragItem)
-      Vue.delete(state.items, ind)
-      await db.moveItem(dragItem, destCategory.id)
-    },
-    async init(store:any){
-      const {state} = store
-      await store.dispatch('loadConfig')
-      state.categories = await db.getCategories()
-      state.selectedCategory = state.categories[0]
-      state.ui = await langSwitch(state.config.lang)
-      thisWindow.show()
-    },
-    async loadConfig({state}){
-      state.configRaw = await db.loadConfig()
-      state.config = state.configRaw.config
-    },
-    async saveConfig({state}){
-      db.saveConfig(state.configRaw)
-    },
-    async langSwitch(store:any, lang:string){
-      store.state.config.lang = lang
-      store.state.ui = await langSwitch(lang)
-      store.dispatch('saveConfig')
-    },
-    async selectTheme(store:any, theme:string){
-      store.state.config.theme = theme
-      store.dispatch('saveConfig')
-      switchTheme(theme)
-    },
-    updateItemsOrder({state}, newOrders:Array<Item>):void{
-      state.items = newOrders
-      newOrders.forEach((o, i)=>{
-        o.sort = i
-        db.items.update(o.id, {'sort':o.sort})
-      })
-    },
-    updateCategoriesOrder({state}, newOrders:Array<Category>):void{
-      state.categories = newOrders
-      newOrders.forEach((o, i)=>{
-        o.sort = i
-        db.categories.update(o.id, {'sort':o.sort})
-      })
-    },
-    updateCategoryName({state}, cate:Category){
-      db.categories.update(cate.id, {name:cate.name})
-    },
-    async addNewCategory({state}, categoryName:string){
-      const cate:Category = await db.addCategory(categoryName)
+    ]
+  })
+  if(targetJsonFiles.filePaths.length === 0){
+    return
+  }
+  const [targetJson] = targetJsonFiles.filePaths
+  if(!existsSync(targetJson)){
+    return eventHub.emit('notify', (state.ui as any).FILE_NOT_FOUND)
+  }
+  try{
+    const jsonTxt = await new Promise<string>(r=>readFile(targetJson, (_e:any, d:any)=>r(d) ))
+    const importData = JSON.parse(jsonTxt)
+    await importData.reduce(async (b:Promise<void>,d:any)=>{
+      await b
+      const {category, items} = d
+      const cate:Category = await db.addCategory(category.name)
+      await items.reduce(async (ib:Promise<void>, i:any)=>{
+        await ib
+        i.cateId = cate.id
+        await db.addItem(i)
+      }, Promise.resolve())
       state.categories.push(cate)
-      state.selectedCategory = cate
-      Vue.nextTick(()=>hub.$emit('adjust'))
-    },
-    async removeCategory({state}, cate:Category){
-      await db.removeCategory(cate)
-      const removeIndex = state.categories.findIndex(c=> c === cate)
-      Vue.delete(state.categories, removeIndex)
-      if(state.selectedCategory === cate){
-        state.items = []
-      }
-      Vue.nextTick(()=>hub.$emit('adjust'))
-    },
-    async getItems({state},cateId:number){
-      const items = await db.getItems(cateId)
-      state.items = items
-      Vue.nextTick(()=>hub.$emit('adjust'))
-    },
-    openHP(){
-      shell.openExternal(globals.HP_URL)
-    },
-    async removeItem({state}, item:Item){
-      await db.removeItem(item)
-      const removeIndex = state.items.findIndex((i:any)=> i === item)
-      Vue.delete(state.items, removeIndex)
-      Vue.nextTick(()=>hub.$emit('adjust'))
-    },
-    async addFile({state}, {filepath}){
-      const cateId = state.selectedCategory.id
-      const items = state.items
-      const fileInfo = await getFileInfo(filepath)
-      fileInfo.cateId = cateId
-      const fileItem = await db.addItem(fileInfo)
-      items.push(fileItem)
-      Vue.nextTick(()=>hub.$emit('adjust'))
-    },
-    async addUrl({state}, {url}){
-      if(!isUrl.test(url)){
-        hub.$emit('notify', state.ui.INVALID_URL)
-        return
-      }
-      const urlItem = await db.addUrlItem(url, state.selectedCategory.id)
-      state.items.push(urlItem)
-      Vue.nextTick(()=>hub.$emit('adjust'))
-    },
-    async copyItemPath(_:void, item:Item){
-      clipboard.writeText(item.path)
-    },
-    async showItemDetail(_:void, item:Item){
-      hub.$emit('show-item-detail', item)
-    },
-    updateItem(_:void, item:Item){
-      db.items.update(item.id, item)
-    },
-    async importJson({state}){
-      const targetJsonFiles = await remote.dialog.showOpenDialog(thisWindow, {
-        title:'Select Myl save data json.',
-        defaultPath:join(remote.app.getPath('desktop'), DEFAULT_JSON_NAME),
-        filters:[
-          {
-            name:'Json',
-            extensions:['json']
-          }
-        ]
-      })
-      if(targetJsonFiles.filePaths.length === 0){
-        return
-      }
-      const [targetJson] = targetJsonFiles.filePaths
-      if(!existsSync(targetJson)){
-        return eventHub.$emit('notify', state.ui.FILE_NOT_FOUND, 'warn')
-      }
-      try{
-        const jsonTxt = await new Promise<string>(r=>readFile(targetJson, 'utf8', (_e:Error, d:any)=>r(d)))
-        const importData = JSON.parse(jsonTxt)
-        await importData.reduce(async (b:Promise<void>,d:any)=>{
-          await b
-          const {category, items} = d
-          const cate:Category = await db.addCategory(category.name)
-          await items.reduce(async (ib:Promise<void>, i:any)=>{
-            await ib
-            i.cateId = cate.id
-            await db.addItem(i)
-          }, Promise.resolve())
-          state.categories.push(cate)
-        }, Promise.resolve())
-        eventHub.$emit('notify', state.ui.IMPORT_DONE)
-      }catch(e){
-        console.error(e)
-        eventHub.$emit('notify', state.ui.FILE_IS_INVALID, 'error')
-      }
-    },
-    async exportJson({state}){
-      const savePath = await remote.dialog.showSaveDialog(thisWindow,{
-        title:'Select save filepath.',
-        defaultPath:join(remote.app.getPath('desktop'), DEFAULT_JSON_NAME)
-      })
-      if(!savePath.filePath){
-        return
-      }
-      const exportData = await db.exportAll(savePath.filePath)
-      await new Promise(r=> writeFile(savePath.filePath, JSON.stringify(exportData), 'utf8' ,r))
-      eventHub.$emit('notify', state.ui.EXPORT_DONE)
-    }
+    }, Promise.resolve())
+    eventHub.emit('notify', (state.ui as any).IMPORT_DONE)
+  }catch(e){
+    console.error(e)
+    eventHub.emit('notify', (state.ui as any).FILE_IS_INVALID)
   }
 }
 
-class MylClass extends Store<object>{
-  constructor(storeData){
-    super(storeData)
-  }
+export function setSize(h:number, w:number){
+  ipcCommunicate('setSize', [h, w])
 }
-const mylClass = new MylClass(storeData)
-mylClass.watch((state:any)=>{
-  return state.config.lang
-},(v,old)=>{
-  mylClass.dispatch('langSwitch', v)
+export function setAlwaysOnTop(tf:boolean){
+  ipcCommunicate('setAlwaysOnTop', tf)
+}
+
+export async function exportJson(){
+  const savePath = await ipcHandleCommunicate('showSaveDialog',{
+    title:'Select save filepath.',
+    defaultPath:join(getDesktopPath(), DEFAULT_JSON_NAME)
+  })
+  if(!savePath.filePath){
+    return
+  }
+  const exportData = await db.exportAll(savePath.filePath)
+  await new Promise(r=> writeFile(savePath.filePath, JSON.stringify(exportData), 'utf8' ,r))
+  eventHub.emit('notify', (state.ui as any).EXPORT_DONE)
+}
+
+watch(state.config.lang as any,(to:string, from:string)=>{
+  langSwitchFn(to as any)
 })
-mylClass.watch((state:any)=>{
-  return state.config.theme
-}, (v,old)=>{
-  mylClass.dispatch('selectTheme', v)
+watch(state.config.theme as any,(to:string, from:string)=>{
+  selectTheme(to)
 })
-export default mylClass
