@@ -1,3 +1,92 @@
+<template>
+  <div @contextmenu="showContentMenu" @dblclick="call" class="item" @dragstart="setDrag">
+    <div class="item-icon" :style="{'background-image':dataUrl}"></div><span class="item-content">{{item.name}}</span>
+  </div>
+</template>
+<script lang="ts">
+import { defineComponent, toRaw } from 'vue'
+import hub from '../ts/event-hub'
+import Item from '../ts/models/item'
+import { ipcRenderer} from 'electron'
+import {Menu, MenuItem} from 'electron'
+import { copyItemPath, removeItem, setDragItem, showItemDetail, state } from '../ts/store'
+export default defineComponent({
+  props:{
+    item:{
+      type:Item,
+      required:true
+    }
+  }, 
+  computed:{
+    sortMode(){
+      return state.sortMode
+    },
+    dataUrl(){
+      return `url(${this.item.icon})`
+    }
+  },
+  methods:{
+    call(){
+      this.item.call()
+    },
+    showContentMenu(ev:MouseEvent){
+      contextMenu(ev, this.item)
+    },
+    setDrag(ev:DragEvent){
+      setDragItem(this.item)
+      const dataTransfer = ev.dataTransfer as DataTransfer
+      if(this.sortMode){
+        dataTransfer.setData('text/plain', this.item.path)
+        dataTransfer.setData('myl/item', '1')
+        return
+      }
+      if(this.item.type === 'url'){
+        ev.stopPropagation()
+        dataTransfer.setData('myl/item', '1')
+        dataTransfer.setData('text/uri-list', this.item.path)
+        return
+      }
+      ev.preventDefault()
+      ev.stopPropagation()
+      ipcRenderer.send('ondragstart', this.item.path)
+    }
+  }
+})
+function contextMenu(ev:MouseEvent, item:Item){
+  const SEND_EVENT_NAME = 'show-item-menu'
+  const RESPONSE_EVENT_NAME = 'select-item-menu'
+  ipcRenderer.once(RESPONSE_EVENT_NAME, (_ev, type)=>{
+    switch(type){
+      case 'openParent':{
+        item.openParent()
+        break
+      }
+      case 'copy':{
+        copyItemPath(item)
+        break
+      }
+      case 'edit':{
+        showItemDetail(item)
+        break
+      }
+      case 'remove':{
+        hub.emit('show-dialog' ,{
+          y:ev.clientY,
+          x:ev.clientX,
+          message:'Remove ok?',
+          onOk(){
+            removeItem(item)
+          },
+          cancelable:true
+        })
+        break
+      }
+    }
+  })
+  ipcRenderer.send(SEND_EVENT_NAME, toRaw(state.ui))
+}
+</script>
+
 <style lang="stylus">
 .item
   padding 0
@@ -29,108 +118,3 @@
     display:inline-block;
     width calc(100% - 23px)
 </style>
-<template>
-  <div @contextmenu="showContentMenu" @dblclick="call" class="item" @dragstart="setDrag">
-    <div class="item-icon" :style="{'background-image':dataUrl}"></div><span class="item-content">{{item.name}}</span>
-  </div>
-</template>
-<script lang="ts">
-import Vue from 'vue'
-import hub from '../ts/event-hub'
-import Item from '../ts/models/item'
-import { ipcRenderer} from 'electron'
-import { mapState } from 'vuex';
-import {Menu, MenuItem} from 'electron'
-export default Vue.extend({
-  props:{
-    item:Item
-  }, 
-  computed:{
-    ...mapState({
-      sortMode:'sortMode'
-    }),
-    dataUrl(){
-      return `url(${this.item.icon})`
-    }
-  },
-  methods:{
-    call(){
-      this.item.call()
-    },
-    showContentMenu(ev:MouseEvent){
-      contextMenu(this.$store, ev, this.item)
-    },
-    setDrag(ev:DragEvent){
-      this.$store.commit('setDragItem', this.item)
-      if(this.sortMode){
-        ev.dataTransfer.setData('text/plain', this.item.path)
-        ev.dataTransfer.setData('myl/item', '1')
-        return
-      }
-      if(this.item.type === 'url'){
-        ev.stopPropagation()
-        ev.dataTransfer.setData('myl/item', '1')
-        ev.dataTransfer.setData('text/uri-list', this.item.path)
-        return
-      }
-      ev.preventDefault()
-      ev.stopPropagation()
-      ipcRenderer.send('ondragstart', this.item.path)
-    }
-  }
-})
-function contextMenu(store, ev:MouseEvent, item:Item){
-  const {state:{ui}} = store
-  const menu = new Menu()
-  const openParent = new MenuItem({
-    id:'OpenParent',
-    accelerator:'o',
-    type:'normal',
-    label:ui.OPEN_PARENT,
-    click(){
-      item.openParent()
-    }
-  })
-  const copyItemPath = new MenuItem({
-    id:'Copy',
-    accelerator:'c',
-    type:'normal',
-    label:ui.COPY,
-    click(){
-      store.dispatch('copyItemPath', item)
-    }
-  })
-  const editItem  = new MenuItem({
-    id:'Edit', 
-    accelerator:'e',
-    type:'normal',
-    label:ui.EDIT, 
-    click(){
-      store.dispatch('showItemDetail', item)
-    }
-  })
-  const removeItem = new MenuItem({
-    id:'Remove',
-    accelerator:'r',
-    type:'normal',
-    label:ui.REMOVE,
-    click(){
-      hub.$emit('show-dialog' ,{
-        y:ev.clientY,
-        x:ev.clientX,
-        message:'Remove ok?',
-        onOk(){
-          store.dispatch('removeItem', item)
-        },
-        cancelable:true
-      })
-    }
-  })
-  menu.append(openParent)
-  menu.append(copyItemPath)
-  menu.append(editItem)
-  menu.append(removeItem)
-  menu.popup({})
-}
-</script>
-
